+++ Assets/Scripts/YangCheng/TalentModule.cs
@@ -14,8 +14,8 @@ namespace My.Player
 
         private int _level = 1;
         // 模拟配置表数据：每升一级加多少属性
-        private Dictionary<int, float> _growthRates;
-        private Dictionary<int, float> _baseStats;
+        private Dictionary<int, long> _growthRates;
+        private Dictionary<int, long> _baseStats;
 
         public EProgressionModule ModuleName => EProgressionModule.Level;
 
@@ -36,8 +36,8 @@ namespace My.Player
             // 逻辑：基础值 + (等级-1 * 成长率)
             foreach (var pair in _baseStats)
             {
-                float growth = _growthRates.ContainsKey(pair.Key) ? _growthRates[pair.Key] : 0;
-                float total = pair.Value + ((_level - 1) * growth);
+                long growth = _growthRates.ContainsKey(pair.Key) ? _growthRates[pair.Key] : 0;
+                long total = pair.Value + ((_level - 1) * growth);
                 targetMap.Add(pair.Key, total);
             }
         }
